Refactor: Move NetworkDataStream data into subclass
* All data used only my Network classes are now in NetworkDataStream
