+++ src/multiplexer.cpp
@@ -30,6 +30,8 @@
 #include "NMEALogWindow.h"
 #include "garmin/jeeps/garmin_wrapper.h"
 #include "OCPN_DataStreamEvent.h"
+#include "datastream.h"
+#include "SerialDataStream.h"
 
 extern PlugInManager    *g_pi_manager;
 extern wxString         g_GPS_Ident;
